Moved engine class to new submodule. Moved custom exception to new submodule
